% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\docType{class}
\name{GEO-class}
\alias{GEO-class}
\alias{GEO}
\title{title: GEO API Client}
\description{
description: Reference class for interacting with the GeoDB API.
}
\section{Methods}{

\describe{
\item{\code{Distance.Between.Places(
  distanceUnit = "KM",
  fromId = NULL,
  fromAddress = NULL,
  fromName = NULL,
  toId = NULL,
  toAddress = NULL,
  toName = NULL
)}}{description: Method used to determine the distance between 2 places\cr
 param - fromId (conditionally required): (character string) the wikidataId of the from place. (Takes priority over fromName and fromAddress, see vignette for more details)\cr
 param - fromName (conditionally required): (character string) the Name of the from place. (Only used if fromId is NULL, see vignette for more details)\cr
 param - fromAddress (conditionally required): (character string) the Name of the from place. (Only used if fromId is NULL, see vignette for more details)\cr
 param - toId (conditionally required): (character string) the wikidataId of the to place. (Takes priority over toName and toAddress, see vignette for more details)\cr
 param - toName (conditionally required): (character string) the Name of the to place. (Only used if toId is NULL, see vignette for more details)\cr
 param - toAddress (conditionally required): (character string) the Name of the to place. (Only used if toId is NULL, see vignette for more details)\cr
 param - distanceUnit (default is KM) : (character string either KM or MI) determines if distances are measured in miles or kilometres.
 return value: The numeric value of the distance between the 2 places.}

\item{\code{FindCities(
  chart = FALSE,
  includeDistricts = FALSE,
  includeCities = TRUE,
  includeAllColumns = FALSE,
  columns = NULL,
  includeDeleted = NULL,
  namePrefix = NULL,
  name = NULL,
  maxPopulation = NULL,
  minPopulation = NULL,
  excludedCountries = NULL,
  includedCountries = NULL,
  excludedCountryIds = NULL,
  includedCountryIds = NULL,
  distanceUnit = "KM",
  radius = NULL,
  longitude = NULL,
  latitude = NULL,
  locationAddress = NULL,
  offset = 0,
  limit = max_limit
)}}{description: Method used to find all Cities and their data using the GEO db api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - includeDistricts : (boolean) Default is false. True if you would like to include districts in the returned places, false otherwise.\cr
param - includeCities : (boolean)Default is true, True if you would like to include cities in the returned places, false otherwise.\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on Cities name\cr
param - name: (character string) Exact City name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of City\cr
param - minPopulation: (numeric integer) filter by min population of City\cr
param - excludedCountries: (character vector) vector of excluded country names (will not include City from these countries)\cr
param - excludedCountryIds: (character vector) vector of excluded country wikiDataId's or codes (will not include City from these countries)\cr
param - includedCountries: (character vector) vector of included country names (only include City from these countries)\cr
param - includedCountryIds: (character vector) vector of included country wikiDataId's or codes (only include City from these countries)\cr
param - distanceUnit: (character string MI or KM) default value is KM, determines if distances are measured in Kilometres or Miles\cr
param - radius: (numeric integer) radius from location to find place, measured in distanceUnit, default is the max radius of your plan\cr
param - longitude: (numeric) longitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - latitude: (numeric) latitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - locationAddress: (character string) the address or name of location in which to find places (only used if longitude or latitude is NULL, see vignette for more details)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindCountries(
  chart = FALSE,
  currencyCode = NULL,
  includeAllColumns = FALSE,
  columns = NULL,
  namePrefix = NULL,
  name = NULL,
  offset = 0,
  limit = max_limit
)}}{description: Method used to find all countries and their data using the GEO db cities api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - currencyCode: (character string) Filter by ISO currency code\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe
param - namePrefix: (character string) Prefix match on country name\cr
param - name: (character string) Exact country name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindDistricts(...)}}{description: Method used to find all Districts and their data using the GEO db api\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on Districts name\cr
param - name: (character string) Exact Districts name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of Districts\cr
param - minPopulation: (numeric integer) filter by min population of Districts\cr
param - excludedCountries: (character vector) vector of excluded country names (will not include Districts from these countries)\cr
param - excludedCountryIds: (character vector) vector of excluded country wikiDataId's or codes (will not include Districts from these countries)\cr
param - includedCountries: (character vector) vector of included country names (only include Districts from these countries)\cr
param - includedCountryIds: (character vector) vector of included country wikiDataId's or codes (only include Districts from these countries)\cr
param - distanceUnit: (character string MI or KM) default value is KM, determines if distances are measured in Kilometres or Miles\cr
param - radius: (numeric integer) radius from location to find place, measured in distanceUnit, default is the max radius of your plan\cr
param - longitude: (numeric) longitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - latitude: (numeric) latitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - locationAddress: (character string) the address or name of location in which to find places (only used if longitude or latitude is NULL, see vignette for more details)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindIslands(...)}}{description: Method used to find all Islands and their data using the GEO db api\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on Island name\cr
param - name: (character string) Exact country name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of islands\cr
param - minPopulation: (numeric integer) filter by min population of islands\cr
param - excludedCountries: (character vector) vector of excluded country names (will not include places from these countries)\cr
param - excludedCountryIds: (character vector) vector of excluded country wikiDataId's or codes (will not include places from these countries)\cr
param - includedCountries: (character vector) vector of included country names (only include places from these countries)\cr
param - includedCountryIds: (character vector) vector of included country wikiDataId's or codes (only include places from these countries)\cr
param - distanceUnit: (character string MI or KM) default value is KM, determines if distances are measured in Kilometres or Miles\cr
param - radius: (numeric integer) radius from location to find place, measured in distanceUnit, default is the max radius of your plan\cr
param - longitude: (numeric) longitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - latitude: (numeric) latitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - locationAddress: (character string) the address or name of location in which to find places (only used if longitude or latitude is NULL, see vignette for more details)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindPlaces(
  chart = FALSE,
  includeDistricts = TRUE,
  includeCities = TRUE,
  includeIslands = TRUE,
  includeAllColumns = FALSE,
  columns = NULL,
  includeDeleted = NULL,
  namePrefix = NULL,
  name = NULL,
  maxPopulation = NULL,
  minPopulation = NULL,
  excludedCountries = NULL,
  includedCountries = NULL,
  excludedCountryIds = NULL,
  includedCountryIds = NULL,
  distanceUnit = "KM",
  radius = NULL,
  longitude = NULL,
  latitude = NULL,
  locationAddress = NULL,
  offset = 0,
  sort = NULL,
  limit = max_limit
)}}{description: Method used to find all places and their data using the GEO db api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - includeDistricts : (boolean) True if you would like to include districts in the returned places, false otherwise.\cr
param - includeCities : (boolean) True if you would like to include cities in the returned places, false otherwise.\cr
param - includeIslands : (boolean) True if you would like to include islands in the returned places, false otherwise.\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on place name\cr
param - name: (character string) Exact place name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of places.\cr
param - minPopulation: (numeric integer) filter by min population of places.\cr
param - excludedCountries: (character vector) vector of excluded country names (will not include places from these countries)\cr
param - excludedCountryIds: (character vector) vector of excluded country wikiDataId's or codes (will not include places from these countries)\cr
param - includedCountries: (character vector) vector of included country names (only include places from these countries)\cr
param - includedCountryIds: (character vector) vector of included country wikiDataId's or codes (only include places from these countries)\cr
param - distanceUnit: (character string MI or KM) default value is KM, determines if distances are measured in Kilometres or Miles\cr
param - radius: (numeric integer) radius from location to find place, measured in distanceUnit, default is the max radius of your plan\cr
param - longitude: (numeric) longitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - latitude: (numeric) latitude value of the nearby location in which to find places (this takes priority over locationAddress, see vignette for more details)\cr
param - locationAddress: (character string) the address or name of location in which to find places (only used if longitude or latitude is NULL, see vignette for more details)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindPlaces.By.CountryAndRegion(
  chart = FALSE,
  includeDistricts = TRUE,
  includeCities = TRUE,
  includeIslands = TRUE,
  includeAllColumns = FALSE,
  columns = NULL,
  country = NULL,
  region = NULL,
  countryId = NULL,
  regionId = NULL,
  includeDeleted = NULL,
  namePrefix = NULL,
  name = NULL,
  maxPopulation = NULL,
  minPopulation = NULL,
  offset = 0,
  limit = max_limit
)}}{description: Method used to find all places in a given country and region using the GEO db api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - country (conditionally required): (character string) name of the given Country (only used if countryId is NULL, see vignette for more details)\cr
param - countryId (conditionally required): (character string) wikidataId or country code of given country (takes priority over country param, see vignette for more details)\cr
param - region (conditionally required): (character string) name of the given Region (only used if regionId is NULL, see vignette for more details)\cr
param - regionId (conditionally required): (character string) wikidataId or region code of given region (takes priority over region param, see vignette for more details)\cr
param - includeDistricts : (boolean) True if you would like to include districts in the returned places, false otherwise.\cr
param - includeCities : (boolean) True if you would like to include cities in the returned places, false otherwise.\cr
param - includeIslands : (boolean) True if you would like to include islands in the returned places, false otherwise.\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on place name\cr
param - name: (character string) Exact place name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of places.\cr
param - minPopulation: (numeric integer) filter by min population of places.\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindPlaces.NearPlace(
  chart = FALSE,
  placeId = NULL,
  placeName = NULL,
  placeAddress = NULL,
  includeDistricts = TRUE,
  includeCities = TRUE,
  includeIslands = TRUE,
  includeAllColumns = FALSE,
  columns = NULL,
  includeDeleted = NULL,
  namePrefix = NULL,
  name = NULL,
  maxPopulation = NULL,
  minPopulation = NULL,
  excludedCountries = NULL,
  includedCountries = NULL,
  excludedCountryIds = NULL,
  includedCountryIds = NULL,
  distanceUnit = "KM",
  radius = NULL,
  offset = 0,
  limit = max_limit
)}}{description: Method used to find all places near a given place and their data using the GEO db api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - charts: (boolean) true to return a chart map, false otherwise\cr
param - placeId (conditionally required): (character string) the wikidataId or native 'id' of the nearby place, (this takes priority over placeName and placeAddress, see vignette for more details)\cr
param - placeName (conditionally required): (character string) the name of the nearby place, (this is only used if placeId is NULL, see vignette for more details)\cr
param - placeAddress (conditionally required): (character string) the address of the nearby place, (this is only used if placeId is NULL, see vignette for more details)\cr
param - includeDistricts : (boolean) True if you would like to include districts in the returned places, false otherwise.\cr
param - includeCities : (boolean) True if you would like to include cities in the returned places, false otherwise.\cr
param - includeIslands : (boolean) True if you would like to include islands in the returned places, false otherwise.\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - includeDeleted: (string either ALL, or NONE, or SINCE_YESTERDAY or SINCE_LAST_WEEK, default is NONE), used to include places that were deleted\cr
param - namePrefix: (character string) Prefix match on country name\cr
param - name: (character string) Exact country name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - maxPopulation: (numeric integer) filter by max population of places.\cr
param - minPopulation: (numeric integer) filter by min population of places.\cr
param - excludedCountries: (character vector) vector of excluded country names (will not include places from these countries)\cr
param - excludedCountryIds: (character vector) vector of excluded country wikiDataId's or codes (will not include places from these countries)\cr
param - includedCountries: (character vector) vector of included country names (only include places from these countries)\cr
param - includedCountryIds: (character vector) vector of included country wikiDataId's or codes (only include places from these countries)\cr
param - distanceUnit: (character string MI or KM) default value is KM, determines if distances are measured in Kilometres or Miles\cr
param - radius: (numeric integer) radius from nearby places to find place, measured in distanceUnit, default is the max radius of your plan\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{FindRegions.ByCountry(
  chart = FALSE,
  includeAllColumns = FALSE,
  columns = NULL,
  country = NULL,
  countryId = NULL,
  limit = max_limit,
  offset = 0,
  name = NULL
)}}{description: Method used to find all Regions and their data in a given country using the GEO db api\cr
param - chart: (boolean) if TRUE render a chart before returning data\cr
param - country (conditionally required): (character string) name of the given Country (only used if countryId is NULL, see vignette for more details)\cr
param - countryId (conditionally required): (character string) wikidataId or country code of given country (takes priority over country param, see vignette for more details)\cr
param - includeAllColumns: (boolean) Logical return all columns\cr
param - columns: (character vector) column names to include in returned dataframe\cr
param - name: (character string) Exact place name (takes priority over 'namePrefix', see vignette for more details.)\cr
param - offset: (numeric integer) Returned data offset, default is 0\cr
param - limit: (numeric integer) Maximum number of results, equal to 'max_limit' field by default\cr
return value: A list with `count` for number of returned records, and `data` for the clean dataframe}

\item{\code{convertLongitudeLatitudeToISO(lat, long)}}{description: Helper method to convert lat and long numeric valies to the ISO lat/long formatted string.\cr
param - lat (required): (numeric) the latitude value\cr
param - long (required): (numeric) the longitude value\cr
return value: returns the ISO formatted string of the long and lat coordinates}

\item{\code{getClosestMatchingCountryNamesOrCodes(countries)}}{description: Helper method to get a list of the corresponding country codes of a given list of country names and/or codes.\cr
param - countries (required): (character vector or character array) the given country codes and/or country names\cr
return value: returns a comma delimited string of the matching country codes, or NULL if none matched}

\item{\code{getCountryNamesAndCodes()}}{description: Helper method to get a dataframe of all country names and country codes.\cr
return value: returns the dataframe with columns: name, code2, and code3}

\item{\code{getLongitudeLatitude(location)}}{description: Helper method to get the ISO formatted long/lat string of a given address or place name.\cr
param - location (required): (character  string) the latitude value\cr
return value: returns the ISO formatted string of the given location}

\item{\code{initialize(api_key = "", max_limit = 10)}}{description: This is the constructor for this class.\cr
 param - api_key (not required) : (character string) Only pass your api key if you are using the pro version, otherwise don't change this parameter and the free version is automatically used.\cr
 param - max_limit : (numeric integer) The default limit value for calling records from the GEO DB api, check  with your plan to see what your maximum is.}

\item{\code{makeRequest(url)}}{description: Helper method to make api request to the given url.\cr
param - url (required): (string) of the url to make the request\cr
return value: returns the response in JSON format}

\item{\code{renderChart(df, title = NULL)}}{description: Helper method to render a quick ggplot chart for returned data frames.\cr
param - df (required): (data.frame) returned data frame\cr
param - title: (character string) plot title\cr
return value: invisible ggplot object, or NULL if it can't be plotted}
}}

